+++ tensorpc/flow/sampleapp/cflow.py
@@ -1,4 +1,4 @@
-from tensorpc.flow import plus
+from tensorpc.flow import plus, mui
 from tensorpc.flow.components.flowplus import ComputeFlow
 from tensorpc.flow import mark_create_layout
 from tensorpc.flow import appctx
@@ -6,7 +6,6 @@ import sys
 from tensorpc import PACKAGE_ROOT
 import numpy as np 
 class ComputeFlowApp:
-
     @mark_create_layout
     def my_layout(self):
         appctx.get_app().set_enable_language_server(True)
@@ -17,7 +16,8 @@ class ComputeFlowApp:
         ]
         self.cflow = ComputeFlow("tensorpc_default_cflow")
         self.panel = plus.InspectPanel({
-            "a": np.zeros((100, 3))
+            "a": np.zeros((100, 3)),
+            "cflow_dev": ComputeFlow("tensorpc_default_cflow_dev"),
         }, use_fast_tree=True, init_layout=self.cflow)
         return self.panel.prop(width="100%", height="100%", overflow="hidden")